Updated code for how long the pump runs based on user input
I had also included a flavor variable. These are hard coded until we figure out how we can update the variable with IoT. There is an issue with sometimes it shows that its pumping for 0 seconds and then repeats the message with the correct time of pumping, it doesn't delay the pump or make it stay on longer, just a bug to figure out.